feat: pass the cart context to redux
