+++ src/components/cart-icon/cart-icon.componet.jsx
@@ -1,14 +1,15 @@
 import React from "react";
 import { CartIconContainer, ItemCount, ShopIcon } from "./cart-icon.style.jsx";
-import { CartCtx } from "../../context/cart.context";
-import { useContext } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { toggleCartDropDown } from "../../store/cart/cart.action.js";
+import { selectCartCount } from "../../store/cart/cart.selector.js";
 
 const CartIcon = () => {
-  const { toggleCartDropdown, cartCount } = useContext(CartCtx);
-  // const count = calculateCartQty(cartItems);
+  const dispatch = useDispatch();
+  const cartCount = useSelector(selectCartCount);
 
   const handleToggleCartDropdown = () => {
-    toggleCartDropdown();
+    dispatch(toggleCartDropDown());
   };
 
   return (
