+++ src/components/nav-bar/nav-bar.component.jsx
@@ -8,16 +8,14 @@ import {
 import { ReactComponent as SamiatLogo } from "../../assets/logo.svg";
 import { useSelector } from "react-redux";
 import { selectCurrentUser } from "../../store/user/user.selector";
-import { CartCtx } from "../../context/cart.context";
-import { useContext } from "react";
 import { signOutUser } from "../../utils/firebase/firebase.util";
 import CartDropdown from "../cart-dropdown/cart-dropdown.component";
 import CartIcon from "../cart-icon/cart-icon.componet";
+import { selectTogleDropdown } from "../../store/cart/cart.selector";
 
 const NavBar = () => {
-  // const { currentUser } = useContext(UserCtx);
   const currentUser = useSelector(selectCurrentUser);
-  const { isCartDropDownOpen } = useContext(CartCtx);
+  const isCartDropDownOpen = useSelector(selectTogleDropdown);
 
   return (
     <NavigationContainer>
